MEDIA: Fix legacy icon of old Mac OS X versions
The icons generated with iconutil is not compatible with old
Mac OS X versions. The only way I have found to generate
compaible icons is to use Icon Composer on an old Mac.
So comment out the rule to generate this icon and update
the icon using one that was generated manually with Icon
Composer.

This fixes bug #11261: MAC OS X: App icon doesn’t display
on OS X 10.5.8.